requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning & Metrics
scikit-learn>=1.3.0

# Progress Bars
tqdm>=4.65.0

grad-cam>=0.1.0
# Optional but recommended for CUDA support
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
# Example: pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu118
